+++ packages/adminui-vue/components/UbRelogin/src/UbAuthNegotiate.vue
@@ -24,7 +24,7 @@
 </template>
 
 <script>
-module.exports = {
+export default {
   name: 'UbAuthNegotiate',
 
   props: {