+++ packages/adminui-vue/components/UbSidebar/init.js
@@ -1,0 +1,13 @@
+/* global BOUNDLED_BY_WEBPACK */
+const Vue = require('vue')
+
+const UbSidebar = require('./src/index.vue').default
+
+module.exports = function () {
+  const domEl = document.createElement('div')
+  document.body.append(domEl)
+
+  new Vue({
+    render: (h) => h(UbSidebar)
+  }).$mount(domEl)
+}