+++ packages/adminui-vue/components/UbTabbar/src/UbTab.vue
@@ -41,12 +41,12 @@ export default {
   padding-right: 28px;
   border-radius: 16px;
   background-color: #fdfdfd;
-  border: solid 1px rgba(var(--dark-blue), 0.54);
+  border: solid 1px rgba(var(--bg-dark), 0.54);
   user-select:none;
   cursor: pointer;
   font-size: 13px;
   font-weight: 600;
-  color: rgba(var(--dark-blue), 0.87);
+  color: rgba(var(--bg-dark), 0.87);
   position: relative;
   white-space:nowrap;
   margin-right: 8px;
@@ -58,12 +58,12 @@ export default {
 }
 
 .ub-tabbar__tab:hover{
-  background-color: rgba(var(--green), 0.16);
-  border-color: rgba(var(--dark-blue), 0.54);
+  background-color: rgba(var(--primary), 0.16);
+  border-color: rgba(var(--bg-dark), 0.54);
 }
 
 .ub-tabbar__tab.active{
-  background-color: rgba(var(--blue), 1);
+  background-color: rgba(var(--bg), 1);
   color: #fff;
   cursor: default;
 }
@@ -115,10 +115,10 @@ export default {
 
 .ub-tabbar__tab-tray.active{
   border-color: #fff;
-  background-color: rgba(var(--dark-blue), 1);
+  background-color: rgba(var(--bg-dark), 1);
 }
 .ub-tabbar__tab-tray:hover{
-  border-color: rgba(var(--green), 1);
-  background-color: rgba(var(--dark-blue), 1);
+  border-color: rgba(var(--primary), 1);
+  background-color: rgba(var(--bg-dark), 1);
 }
 </style>